# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WTR Ghana Water Platform API
  version: 1.0.0
  description: Multi-tenant community water utility management API.
paths:
...
  /api/meters/{id}/:
    get:
      operationId: meters_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this meter.
        required: true
      tags:
      - meters
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meter'
          description: ''
    put:
      operationId: meters_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this meter.
        required: true
      tags:
      - meters
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Meter'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Meter'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Meter'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meter'
          description: ''
    patch:
      operationId: meters_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this meter.
        required: true
      tags:
      - meters
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMeter'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedMeter'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedMeter'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meter'
          description: ''
    delete:
      operationId: meters_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this meter.
        required: true
      tags:
      - meters
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/meters/lookup/:
    get:
      operationId: meters_lookup_retrieve
      description: |-
        Find a meter by the number printed on it (serial number) or its system ID, within the caller's
        community. Returns everything a reader needs to key in a reading: household, previous reading,
        usual consumption, meter status. `?number=`.
      tags:
      - meters
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meter'
          description: ''
...
  /api/readings/:
    get:
      operationId: readings_list
      parameters:
      - in: query
        name: customer
        schema:
          type: integer
      - in: query
        name: is_anomalous
        schema:
          type: boolean
      - in: query
        name: meter
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: read_by
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: source
        schema:
          type: string
          enum:
          - ESTIMATED
          - IOT
          - MANUAL
          - MOBILE
          - OCR
        description: |-
          * `MANUAL` - Manual (office)
          * `MOBILE` - Mobile app
          * `OCR` - Photo OCR
          * `IOT` - Smart meter
          * `ESTIMATED` - Estimated
      - in: query
        name: status
        schema:
          type: string
          enum:
          - BILLED
          - PENDING
          - REJECTED
          - VALIDATED
        description: |-
          * `PENDING` - Pending validation
          * `VALIDATED` - Validated
          * `REJECTED` - Rejected
          * `BILLED` - Billed
      tags:
      - readings
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMeterReadingList'
          description: ''
    post:
      operationId: readings_create
      tags:
      - readings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeterReading'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MeterReading'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MeterReading'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterReading'
          description: ''
  /api/readings/{id}/:
    get:
      operationId: readings_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this meter reading.
        required: true
      tags:
      - readings
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterReading'
          description: ''
    put:
      operationId: readings_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this meter reading.
        required: true
      tags:
      - readings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeterReading'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MeterReading'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MeterReading'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterReading'
          description: ''
    patch:
      operationId: readings_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this meter reading.
        required: true
      tags:
      - readings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMeterReading'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedMeterReading'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedMeterReading'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterReading'
          description: ''
    delete:
      operationId: readings_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this meter reading.
        required: true
      tags:
      - readings
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    ConditionEnum:
      enum:
      - GOOD
      - FAIR
      - POOR
      - DAMAGED
      type: string
      description: |-
        * `GOOD` - Good
        * `FAIR` - Fair
        * `POOR` - Poor
        * `DAMAGED` - Damaged
...
    Meter:
      type: object
      description: '`serial_number` is the number printed on the meter — the one a
        reader keys in. It is mandatory and unique.'
      properties:
        id:
          type: integer
          readOnly: true
        serial_number:
          type: string
          maxLength: 80
        customer_name:
          type: string
          readOnly: true
        customer_code:
          type: string
          readOnly: true
        property_label:
          type: string
          readOnly: true
        replaced_by_id:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        meter_id:
          type: string
          readOnly: true
        meter_type:
          type: string
          maxLength: 60
        meter_size:
          type: string
          description: e.g. 15mm, 1/2"
          maxLength: 30
        manufacturer:
          type: string
          maxLength: 100
        installation_date:
          type: string
          format: date
          nullable: true
        installation_location:
          type: string
          maxLength: 200
        initial_reading:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
        current_reading:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        status:
          $ref: '#/components/schemas/MeterStatusEnum'
        condition:
          $ref: '#/components/schemas/ConditionEnum'
        last_inspection:
          type: string
          format: date
          nullable: true
        last_maintenance:
          type: string
          format: date
          nullable: true
        is_smart:
          type: boolean
        iot_device_id:
          type: string
          maxLength: 120
        notes:
          type: string
        community:
          type: integer
          readOnly: true
        customer:
          type: integer
          nullable: true
        property:
          type: integer
          nullable: true
        replaced_by:
          type: integer
          readOnly: true
          nullable: true
      required:
      - community
      - created_at
      - current_reading
      - customer_code
      - customer_name
      - id
      - meter_id
      - property_label
      - replaced_by
      - replaced_by_id
      - serial_number
      - updated_at
    MeterReading:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        meter_code:
          type: string
          readOnly: true
        customer_name:
          type: string
          readOnly: true
        customer_code:
          type: string
          readOnly: true
        read_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        reading_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
        previous_reading:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        consumption:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        reading_date:
          type: string
          format: date
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        gps_distance_m:
          type: integer
          readOnly: true
          nullable: true
          description: Distance between reader GPS and customer GPS
        photo:
          type: string
          format: uri
          nullable: true
        photo_url:
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        device_id:
          type: string
          maxLength: 120
        client_reading_id:
          type: string
          nullable: true
          description: Client-generated UUID for offline sync idempotency
          maxLength: 80
        source:
          $ref: '#/components/schemas/MeterReadingSourceEnum'
        status:
          allOf:
          - $ref: '#/components/schemas/MeterReadingStatusEnum'
          readOnly: true
        is_anomalous:
          type: boolean
          readOnly: true
        anomaly_flags:
          readOnly: true
        anomaly_note:
          type: string
          readOnly: true
        ocr_detected_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          nullable: true
        validated_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        notes:
          type: string
        community:
          type: integer
          readOnly: true
        meter:
          type: integer
        customer:
          type: integer
          readOnly: true
          nullable: true
        read_by:
          type: integer
          readOnly: true
          nullable: true
        validated_by:
          type: integer
          readOnly: true
          nullable: true
      required:
      - anomaly_flags
      - anomaly_note
      - community
      - consumption
      - created_at
      - customer
      - customer_code
      - customer_name
      - gps_distance_m
      - id
      - is_anomalous
      - meter
      - meter_code
      - previous_reading
      - read_by
      - read_by_name
      - reading_date
      - reading_value
      - status
      - updated_at
      - validated_at
      - validated_by
    MeterReadingSourceEnum:
      enum:
      - MANUAL
      - MOBILE
      - OCR
      - IOT
      - ESTIMATED
      type: string
      description: |-
        * `MANUAL` - Manual (office)
        * `MOBILE` - Mobile app
        * `OCR` - Photo OCR
        * `IOT` - Smart meter
        * `ESTIMATED` - Estimated
    MeterReadingStatusEnum:
      enum:
      - PENDING
      - VALIDATED
      - REJECTED
      - BILLED
      type: string
      description: |-
        * `PENDING` - Pending validation
        * `VALIDATED` - Validated
        * `REJECTED` - Rejected
        * `BILLED` - Billed
...
    MeterStatusEnum:
      enum:
      - AVAILABLE
      - INSTALLED
      - ACTIVE
      - FAULTY
      - BLOCKED
      - REMOVED
      - REPLACED
      - RETIRED
      type: string
      description: |-
        * `AVAILABLE` - Available (in stock)
        * `INSTALLED` - Installed (not yet active)
        * `ACTIVE` - Active
        * `FAULTY` - Faulty
        * `BLOCKED` - Blocked
        * `REMOVED` - Removed
        * `REPLACED` - Replaced
        * `RETIRED` - Retired
...
    PaginatedMeterReadingList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/MeterReading'
...
    PatchedMeter:
      type: object
      description: '`serial_number` is the number printed on the meter — the one a
        reader keys in. It is mandatory and unique.'
      properties:
        id:
          type: integer
          readOnly: true
        serial_number:
          type: string
          maxLength: 80
        customer_name:
          type: string
          readOnly: true
        customer_code:
          type: string
          readOnly: true
        property_label:
          type: string
          readOnly: true
        replaced_by_id:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        meter_id:
          type: string
          readOnly: true
        meter_type:
          type: string
          maxLength: 60
        meter_size:
          type: string
          description: e.g. 15mm, 1/2"
          maxLength: 30
        manufacturer:
          type: string
          maxLength: 100
        installation_date:
          type: string
          format: date
          nullable: true
        installation_location:
          type: string
          maxLength: 200
        initial_reading:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
        current_reading:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        status:
          $ref: '#/components/schemas/MeterStatusEnum'
        condition:
          $ref: '#/components/schemas/ConditionEnum'
        last_inspection:
          type: string
          format: date
          nullable: true
        last_maintenance:
          type: string
          format: date
          nullable: true
        is_smart:
          type: boolean
        iot_device_id:
          type: string
          maxLength: 120
        notes:
          type: string
        community:
          type: integer
          readOnly: true
        customer:
          type: integer
          nullable: true
        property:
          type: integer
          nullable: true
        replaced_by:
          type: integer
          readOnly: true
          nullable: true
    PatchedMeterReading:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        meter_code:
          type: string
          readOnly: true
        customer_name:
          type: string
          readOnly: true
        customer_code:
          type: string
          readOnly: true
        read_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        reading_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
        previous_reading:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        consumption:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        reading_date:
          type: string
          format: date
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        gps_distance_m:
          type: integer
          readOnly: true
          nullable: true
          description: Distance between reader GPS and customer GPS
        photo:
          type: string
          format: uri
          nullable: true
        photo_url:
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        device_id:
          type: string
          maxLength: 120
        client_reading_id:
          type: string
          nullable: true
          description: Client-generated UUID for offline sync idempotency
          maxLength: 80
        source:
          $ref: '#/components/schemas/MeterReadingSourceEnum'
        status:
          allOf:
          - $ref: '#/components/schemas/MeterReadingStatusEnum'
          readOnly: true
        is_anomalous:
          type: boolean
          readOnly: true
        anomaly_flags:
          readOnly: true
        anomaly_note:
          type: string
          readOnly: true
        ocr_detected_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          nullable: true
        validated_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        notes:
          type: string
        community:
          type: integer
          readOnly: true
        meter:
          type: integer
        customer:
          type: integer
          readOnly: true
          nullable: true
        read_by:
          type: integer
          readOnly: true
          nullable: true
        validated_by:
          type: integer
          readOnly: true
          nullable: true
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT